File Refactoring and Defaults Editing
This commit removes some template files.

- Content file refactoring.
- Added Ravine map.
- Two Ravine submaps are included in this commit but are not used. They
will be removed in the next commit.
- Added FValorHeroInitializationProperties struct to store some
initialization properties, similar to ValorCharacterStatContainer. These
exist to make replication of stats easier and reduces the number of
network requests for replication.
- Added UValorUserWidget parent class for all future UUserWidgets to
extend from. The intent is to reduce editor autocomplete cluttering and
allow some code implementation of potentially critical sections.
- Added EValorTeam::Invalid as a hidden enum value. Even though UENUM
with flags will assign the first value 1, the same can not be said of
native usage. None will be assigned a value of 0U which can't be used in
flag operations.