Camera, GUI, and Input Work
This commit implements some rough camera, GUI, and input functionality.

- Created ValorLevelScriptActor that extends LevelScriptActor. This is
the same as a Level Blueprint. DefaultEngine.ini was updated to reflect
this change.
- Added the keybindings. DefaultInput.ini was updated to reflect this
change.
- Updated W_MainGameGUI_Natsu to call the internal methods (more on that
in a bit). This file name will be changed later.
- Removed unused sublevels.
- Added ValorMainInterfaceWidget as the parent class for the main user
interface.
- ValorUserWidget file moved to GUI subfolder.
- Removed unused methods in ValorHeroCharacter_Natsu to clean up the
VTable.
- Removed unused methods in ValorHeroAIController to clean up the
VTable.
- Removed unused methods in ValorHeroCharacter to clean up the VTable.
- Added mouse functionality to ValorHeroCharacterProxy that will pan the
camera if the mouse gets near the window's edges.
- Updated the camera boom to the proper orientation in
ValorHeroCharacterProxy.cpp.
- Removed unused methods in ValorCharacter to clean up the VTable.
- Removed unused methods in ValorPlayerController to clean up the
VTable.
- Overrode GameViewportClient::ReceivedFocus to fire a delegate when
called that other classes can subscribe to. Due to a bug in the engine,
however, this only fires on widgets and not the actual game window.
- Moved ValorHUD to GUI subfolder.
- Switched references from ValorUserWidget to ValorMainInterfaceWidget
in ValorHeroInitializationProperties in order to further restrict
available classes and better indicate intent.
- Added measure to ensure that the server's Character and AIController
have the same PlayerState as the PlayerController's PlayerState
instance. To make this work some timers were added due to issues in the
order of BeginPlay between controllers and pawns.
- Added Valor ClassGroup to ValorPlayerStart.