Finished StatComponent
After some health issues I am back with the next commit. This commit
sets the StatComponent components into a usable state.

- Changed HeroStats.xlsx to HeroStats.xlsm with an embedded macro to
export all selected sheets as csv.
- Removed the "Hero" prefix from stat data.
- Updated Natsu's GUI component with a test health and mana bar to test
stat replication and ease of data access.
- Updated old datatable references to new files.
- Added ValorUnitStatComponent to separate generic actor attachments and
unit attachments. ValorUnitStatComponent tracks mana, movement speed,
experience, level, and cooldown reduction. ValorStatComponent tracks
health.
- Added ValorHeroStatComponent that extends ValorUnitStatComponent that
overrides the level and experience calculations to retrieve the data
from the PlayerState.
- Changed the initialization method of StatComponents to parse the raw
row data instead of attempting to map and read them from variables. This
makes the code cleaner and allows entire rows to be parsed as a single
array. This also means I can do custom error handling, so missing rows
will not crash the application but instead trigger an ensure if an
attempt was made to access a property that requires the row.
- Removed StatComponent TickComponent inclusion as the components will
never tick outside of the custom timer.
- Getters in StatComponents now require an EValorStatType parameter to
select from Base, Current, Bonus, and Max. This makes it easier to
maintain the code as all info related to a certain stat is in one place.
- Added replication code to StatComponent and subclasses.
- Added DataBlueprintLibrary with 3 different methods that parse an
array of strings into an array of floats, int32s, and uint8s.
- Added InputBlueprintLibrary that exposes runtime keybinding
functionality.
- Added ValorActionInputBinding and ValorAxisInputBinding struct
wrappers to expose input structs to blueprints.
- Removed the VALOR_PLAYER_MAX_LEVEL macro and added GetMaximumPlayers()
and GetMaximumLevel() methods in AValorGameMode.
- Removed ValorCharacterExperienceTableRow struct since it is no longer
required with the new data reading method.
- Renamed ValorCharacterStatTableRow to ValorStatTableRow.
- Added expected row names in ValorDataDefinitions.h so they are not
hard-coded.
- Added AdjustExperience() method to ValorPlayerState.
- Exposed Character variable of ValorHeroCharacterProxy to blueprints as
a readonly property so the GUI can grab the player's stats.