VG-15 #close Added IValorSpawnable
The IValorSpawnable interface can now be attached to any class
(recommended root class of common classes). This exposes a Spawn()
method that is used to initialize the character. Due to this Initialize
is now protected.

Spawn() takes care of controller possession (with the exception of
player characters - this will change after VG-7) as well.

Spawn() has two parameters, an APlayerState and a FValorVariantData or a
pointer. The former is used when a PlayerState needs to be associated
with an AI character (as with a proxy or an AI hero character), and the
latter two are used in case extra data needs to be passed to Spawn() in
order to facilitate proper spawning behavior.

I also moved all interfaces to a single directory (and subdirectories).