VG-15 #close Core Class Restructure
Made IValorClickableInterface, IValorHoverableInterface, and
IValorUnitInterface. Attached them to AValorCharacter (AValorStructure
pending). Actors are automatically hooked to these interfaces at the
higher level, but lower level classes will need to override if more
functionality is required.

The detection methods in PlayerController are not required as there are
detection delegates attached to components. These interfaces are bound
to the root CapsuleComponent.