VG-18 #resolve Unit Right Clicking
Right-clicking on an enemy unit now moves the unit into attack range and
sets them up to attack. A ValorAIDefinitions class was added that
defines the blackboard key names so they aren't hard-coded.

I also fixed a lot of bugs regarding interfaces. It turns out that
things bug out when UFUNCTION specifiers do not match in the override.
If a method is defined BlueprintCallable in the interface it must be
defined as BlueprintCallable in the override or it will return bad data.
I was getting compiler errors that BlueprintPure is not allowed in
interfaces, but I'll try again later. It seems like the editor knows
that these are pure regardless as they don't require an execution pin
even with BlueprintCallable.